requirements.txt
sentence-transformers>=2.6.1
langchain
streamlit
pypdf
python-dotenv
pinecone[grpc]
langchain_pinecone
langchain_community
langchain_openai
langchain_experimental
torch
torchvision
matplotlib
pillow
torchxrayvision

-e .